+++ webpack.config.js
@@ -1,9 +1,9 @@
 const del = require('del');
 const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+// const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const magicSassImporter = require('node-sass-magic-importer');
-const SassLintPlugin = require('sasslint-webpack-plugin');
+// const SassLintPlugin = require('sasslint-webpack-plugin');
 const babelOptions = require('./.babelrc.json');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const WebpackOnBuildPlugin = require('on-build-webpack');
@@ -17,12 +17,12 @@ const cleanDistBeforeBuild = new CleanWebpackPlugin('dist', {
     root: process.cwd()
 });
 
-const sassLinter = new SassLintPlugin({
-    configFile: path.join(__dirname, '.sass-lint.yml'),
-    glob: '!(node_modules/)**/*.scss',
-    failOnError: true,
-    ignorePlugins: ['extract-text-webpack-plugin']
-});
+// const sassLinter = new SassLintPlugin({
+//     configFile: path.join(__dirname, '.sass-lint.yml'),
+//     glob: '!(node_modules/)**/*.scss',
+//     failOnError: true,
+//     ignorePlugins: ['extract-text-webpack-plugin']
+// });
 
 // This will extract the styles from the bundle.js file.
 const extractCss = new MiniCssExtractPlugin({ filename: '[name].bundle.css' });
@@ -125,7 +125,7 @@ module.exports = {
 
     plugins: [
         cleanDistBeforeBuild,
-        sassLinter,
+        //        sassLinter,
         extractCss,
         // new OptimizeCSSAssetsPlugin(), // if you optimise, you lose the css source map
         cleanUpThemeJsFiles